+++ src/main/java/cn/zjut/hotel/service/HotelOrderServiceInterface.java
@@ -17,5 +17,5 @@ public interface HotelOrderServiceInterface {
 
     boolean modifyOneOrderById(HotelOrder order);
 
-    boolean removeOneOrderById(int orderId);
+    boolean removeOneOrderById(long orderId);
 }
